# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个简单的校园点评网站的后端 API 文档。",
        "title": "杭电点评 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/reviews/pending": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取等待管理员审核的点评列表，支持分页、搜索和排序。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "待审核点评列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "-created_at",
                        "description": "排序字段 (e.g., -created_at, rating)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ReviewListResult"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            }
        },
        "/admin/reviews/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "永久删除指定 ID 的点评及其关联的图片。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "删除点评",
                "parameters": [
                    {
                        "type": "string",
                        "description": "点评 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "删除成功"
                    },
                    "400": {
                        "description": "无效的点评 ID",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "404": {
                        "description": "点评不存在",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            }
        },
        "/admin/reviews/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "批准或拒绝一个点评。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "更新点评状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "点评 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "状态更新请求 (status: 'approved' 或 'rejected')",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "reason": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/models.Review"
                        }
                    },
                    "400": {
                        "description": "无效的请求",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "404": {
                        "description": "点评不存在",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "409": {
                        "description": "点评状态已被处理，无法再次修改",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AuthorResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "dto.ImageResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "dto.ReviewResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/dto.AuthorResponse"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ImageResponse"
                    }
                },
                "rating": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.Review": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/models.User"
                },
                "author_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ReviewImage"
                    }
                },
                "rating": {
                    "type": "number"
                },
                "rejection_reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.ReviewStatus"
                },
                "store": {
                    "$ref": "#/definitions/models.Store"
                },
                "store_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ReviewImage": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "review_id": {
                    "type": "string"
                },
                "storage_key": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.ReviewStatus": {
            "type": "string",
            "enum": [
                "pending",
                "approved",
                "rejected"
            ],
            "x-enum-varnames": [
                "ReviewStatusPending",
                "ReviewStatusApproved",
                "ReviewStatusRejected"
            ]
        },
        "models.Store": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "auto_created": {
                    "type": "boolean"
                },
                "average_rating": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "rejection_reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.StoreStatus"
                },
                "total_reviews": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.StoreStatus": {
            "type": "string",
            "enum": [
                "pending",
                "approved",
                "rejected"
            ],
            "x-enum-varnames": [
                "StoreStatusPending",
                "StoreStatusApproved",
                "StoreStatusRejected"
            ]
        },
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "problem.Details": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "services.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "services.ReviewListResult": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ReviewResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/services.Pagination"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}